requirements/development.txt
-r base.txt
pytest
pytest-watch
pytest-django
backports.shutil-get-terminal-size
coverage
decorator
ipdb
ipython
ipython-genutils
traitlets
pexpect
pickleshare
ptyprocess
py
simplegeneric
mixer==5.6.6
mock
flake8
django-seed

#important
#use this versions
pytest-cov
fake-factory==0.5.7


django-debug-toolbar
beautifulsoup4
lxml
numpy
pandas
selenium
django-webpack-loader
pytest-ordering
pytest-sugar
django-debug-toolbar-template-profiler
django-debug-toolbar-force